+++ R/convertWarnings.R
@@ -1,12 +1,14 @@
 #' Wrapper to convert all/specific warning messages to errors
 #'
-#' Function is intended to be a passive wrapper to functions
+#' Function is intended to be a passive wrapper/catch-all to functions
 #' that are known to throw warning messages that should generally
 #' be treated as errors instead (e.g., non-positive definite matrix
 #' warnings, negative variance estimate warnings, etc). Specific
 #' warning messages can be caught if specified, otherwise all
 #' detected warning messages will be converted to errors for the evaluated
-#' R expression.
+#' R expression. R expressions may be evaluated in isolation for greater
+#' specificity or evaluated after the function(s) terminate for more general
+#' warning detection behaviour.
 #'
 #' General goal of this function is to \emph{explicitly}
 #' indicate warning that are problematic. In many function implementations
@@ -18,7 +20,11 @@
 #' as well as \code{option(warn=2)} to convert all warnings to errors
 #' globally).
 #'
-#' @param expr expression to be evaluated (e.g., \code{ret <- myfun(args)})
+#' @param expr expression to be evaluated (e.g., \code{ret <- myfun(args)}).
+#'   Omitting this argument will use the result from the latest
+#'   \code{\link{warning}} call, which has a greater risk of throwing false
+#'   positives (particularly if \code{warning2error} left unspecified)
+#'   and has less specificity, but can be more convenient
 #'
 #' @param warning2error a character vector of warning messages
 #'   that should be converted to errors. Each warning message is
@@ -97,19 +103,48 @@
 #'                        c("Show this warning", "Show a different warning"))
 #' ret3
 #'
+#'
+#' #####################
+#' # convertWarnings() can be evaluated after function calls
+#' #   (temporal result of warnings() caught and tested). Provides less
+#' #   specificity and higher likely of false positives and negatives, but
+#' #   can be more convenient
+#'
+#' fun(warn1=TRUE)     # warning raised
+#' convertWarnings()   # all warnings translated to error
+#' fun(warn2=TRUE)
+#' convertWarnings()   # initial warn1=TRUE does not carry over!
+#'
+#' # multiple warnings raised, though converted to errors given specificity
+#' fun(warn1=TRUE, warn2=TRUE)
+#' convertWarnings(warning2error=c('Only error with this warning')) # no errors
+#' convertWarnings(warning2error=c('Show a different warning')) # error
+#'
 #' }
 #'
 convertWarnings <- function(expr, warning2error=NULL, muffle=FALSE){
-    ret <- withCallingHandlers({
-        eval(expr)
-    }, warning=function(w) {
-        message <- conditionMessage(w)
+
+    stopit <- function(message, warning2error){
         if(is.null(warning2error)) stop(message, call.=FALSE)
         sapply(warning2error, function(warn){
             match <- grepl(warn, message)
             if(match) stop(message, call.=FALSE)
         })
-        if(muffle) invokeRestart("muffleWarning")
-    })
+    }
+
+    if(missing(expr)){
+        message <- names(warnings())
+        if(length(message))
+            sapply(message, stopit, warning2error=warning2error)
+        return(invisible(NULL))
+    } else {
+        ret <- withCallingHandlers({
+            eval(expr)
+        }, warning=function(w) {
+            message <- conditionMessage(w)
+            stopit(message, warning2error = warning2error)
+            if(muffle) invokeRestart("muffleWarning")
+        })
+    }
     ret
 }